+++ java-diff-utils/src/test/java/com/github/difflib/patch/PatchTest.java
@@ -1,7 +1,6 @@
 package com.github.difflib.patch;
 
 import com.github.difflib.DiffUtils;
-import com.github.difflib.algorithm.DiffException;
 import java.util.Arrays;
 import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -11,7 +10,7 @@ import org.junit.jupiter.api.Test;
 public class PatchTest {
 
     @Test
-    public void testPatch_Insert() throws DiffException {
+    public void testPatch_Insert() {
         final List<String> insertTest_from = Arrays.asList("hhh");
         final List<String> insertTest_to = Arrays.asList("hhh", "jjj", "kkk", "lll");
 
@@ -24,7 +23,7 @@ public class PatchTest {
     }
 
     @Test
-    public void testPatch_Delete() throws DiffException {
+    public void testPatch_Delete() {
         final List<String> deleteTest_from = Arrays.asList("ddd", "fff", "ggg", "hhh");
         final List<String> deleteTest_to = Arrays.asList("ggg");
 
@@ -37,7 +36,7 @@ public class PatchTest {
     }
 
     @Test
-    public void testPatch_Change() throws DiffException {
+    public void testPatch_Change() {
         final List<String> changeTest_from = Arrays.asList("aaa", "bbb", "ccc", "ddd");
         final List<String> changeTest_to = Arrays.asList("aaa", "bxb", "cxc", "ddd");